Adds District class with name attribute
