+++ lib/district.rb
@@ -1,0 +1,8 @@
+class District
+
+  attr_reader :name
+  
+  def initialize(data)
+    @name = data[:name]
+  end
+end
