+++ lib/district_repository.rb
@@ -1,0 +1,9 @@
+require 'csv'
+
+class DistrictRepository
+
+  def load_data(file)
+    data = CSV.open(file, headers: true, header_converters: :symbol)
+  end
+
+end
