--- test/districtrepositorytest.rb
@@ -1,12 +1,0 @@
-require './test/test_helper'
-require './lib/districtrepository'
-
-class DistrictRepositoryTest < Minitest::Test
-
-def test_it_exists
-  dr = DistrictRepository.new
-
-  assert_instance_of DistrictRepository, dr
-end
-
-end
